+++ src/app/profile/page.tsx
@@ -4,9 +4,9 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { User, Mail, LayoutDashboard } from "lucide-react";
-import { ProfileForm } from "@/components/ProfileForm";
-import { ExportLinksButton } from "@/components/ExportLinksButton";
-import { DeleteAccountButton } from "@/components/DeleteAccountButton";
+import { ProfileForm } from "@/components/profile/ProfileForm";
+import { ExportLinksButton } from "@/components/dashboard/ExportLinksButton";
+import { DeleteAccountButton } from "@/components/profile/DeleteAccountButton";
 import Link from "next/link";
 import { Button } from "@/components/ui/button";
 import Image from "next/image";
